requirements.txt
Django==1.6.8
Django-Select2==4.2.2
Pillow==2.6.1
South==1.0
Unidecode==0.04.16
aldryn-common==0.0.6
argparse==1.2.1
dj-database-url==0.3.0
django-appconf==0.6
django-appdata==0.1.3
django-classy-tags==0.5.1
django-cms==3.0.6
django-filer==0.9.7
django-hvad==0.5.0
django-mptt==0.6.1
django-polymorphic==0.6
django-reversion==1.8.4
django-sekizai==0.7
django-taggit==0.11.2
djangocms-admin-style==0.2.2
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.5.4
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.4
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.0
djangocms-video==0.1
easy-thumbnails==2.2
html5lib==0.999
psycopg2==2.5.4
pytz==2014.7
simplejson==3.6.5
six==1.8.0
wsgiref==0.1.2